requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-modeladmin-reorder==0.3.1
djangorestframework==3.14.0
ElasticEmail==4.0.23
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
